feat: delete user

+++ app/Http/Controllers/Admin/UserController.php
@@ -10,6 +10,7 @@ use Illuminate\Http\Request;
 use Yajra\DataTables\Facades\DataTables;
 use App\Constants;
 use App\Exceptions\NotFoundError;
+use App\Models\Identifier;
 use App\Models\ModelHasRole;
 use App\Models\Role;
 use App\Models\Unit;
@@ -29,11 +30,11 @@ class UserController extends Controller
     public function index()
     {
         $userStatus = $this->constants->user_status;
-        $roles = Role::all();
-        $units = Unit::all();
+
+        $identifiers = Identifier::all();
 
         return view('admin.user.index', compact([
-            'userStatus', 'roles', 'units'
+            'userStatus', 'identifiers'
         ]));
     }
 
@@ -71,8 +72,7 @@ class UserController extends Controller
                 'email' => 'required|string|unique:users,email',
                 'password' => 'required|string|min:8',
                 'status' => ['nullable', Rule::in($this->constants->user_status)],
-                'role_id' => 'nullable|exists:roles,id',
-                'unit_id' => 'nullable|exists:units,id',
+                'identifiers' => 'required|array|min:1|exists:identifiers,id',
                 'signature' => 'string|nullable',
                 'avatar' => 'string|nullable',
             ]);
@@ -93,11 +93,7 @@ class UserController extends Controller
                 "status" => $request->status
             ]);
 
-            ModelHasRole::create([
-                "role_id" => $request->role_id,
-                "model_type" => "App\Models\User",
-                "model_id" => $user->id
-            ]);
+            $user->identifiers()->attach($request->identifiers);
 
             return response()->json([
                 "status" => "success",
@@ -151,10 +147,17 @@ class UserController extends Controller
         }
     }
 
-    public function destroy(string $id)
+    public function destroy(Request $request)
     {
         try {
-            $user = User::findOrFail($id);
+            $user = User::findOrFail($request->id);
+
+
+            if (!$user) {
+                throw new NotFoundError("User tidak ditemukan");
+            }
+
+            $user->identifiers()->detach();
 
             $user->delete();
